+++ src/phoenix/server/api/routers/__init__.py
@@ -1,6 +1,6 @@
 from .auth import router as auth_router
 from .embeddings import create_embeddings_router
-from .oauth import router as oauth2_router
+from .oauth2 import router as oauth2_router
 from .v1 import create_v1_router
 
 __all__ = [
